unauthorized users can't accesses scripts

+++ delete-product-processing.php
@@ -1,11 +1,22 @@
 <?php
 
 session_start();
+
+if (!isset($_SESSION['admin_logged_in'])) {
+    header('Location: index.php');
+    die();    
+}
+
 require_once 'common.php';
 
 $id = isset($_POST['productId']) ? strip_tags($_POST['productId']) : '';
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST' && filter_var($id, FILTER_VALIDATE_INT)) {
+
+    $query = "DELETE FROM ordersproducts WHERE product_id = :id;";
+    $stmt = $pdo->prepare(query: $query);
+    $stmt->bindParam(':id', $id);
+    $stmt->execute();
 
     $query = "DELETE FROM products WHERE id = :id;";
     $stmt = $pdo->prepare(query: $query);
